add: auth file under router for auth purposes
