+++ src/server/api/root.ts
@@ -1,5 +1,6 @@
 import { createTRPCRouter } from "@/src/server/api/trpc";
 import { wordRouter } from "./routers/word";
+import { authRouter } from "./routers/auth";
 
 /**
  * This is the primary router for your server.
@@ -8,6 +9,7 @@ import { wordRouter } from "./routers/word";
  */
 export const appRouter = createTRPCRouter({
   word: wordRouter,
+  auth: authRouter,
 });
 
 // export type definition of API
